--- src/test/java/searchengine/controllers/IndexControllerTests.java
@@ -1,29 +1,0 @@
-package searchengine.controllers;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@Disabled
-class IndexControllerTests {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    void startIndexing() {
-    }
-
-    @Test
-    void stopIndexing() {
-    }
-
-    @Test
-    void indexPage() {
-    }
-}
