--- src/test/java/searchengine/controllers/SearchControllerTests.java
@@ -1,23 +1,0 @@
-package searchengine.controllers;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@Disabled
-class SearchControllerTests {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    void search() {
-    }
-}
