--- src/test/java/searchengine/repository/IndexRepositoryTest.java
@@ -1,7 +1,0 @@
-package searchengine.repository;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class IndexRepositoryTest {
-
-}
