--- src/test/java/searchengine/repository/SiteRepositoryTest.java
@@ -1,7 +1,0 @@
-package searchengine.repository;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class SiteRepositoryTest {
-
-}
