--- src/test/resources/application-test.yml
@@ -1,20 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/search_engine?createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true&useSSL=false&useUnicode=true&characterEncoding=UTF-8&serverTimezone=UTC&stringtype=unspecified}
-    username: root
-    password: testtest
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-
-server:
-  port: 8080
-
-indexing-settings:
-  sites:
-    - url: https://www.lenta.ru
-      name: Лента.ру
-    - url: https://www.skillbox.ru
-      name: Skillbox
-    - url: https://www.playback.ru
-      name: PlayBack.Ru
